bots/requirements.txt
python-telegram-bot==21.3
discord.py==2.3.2
aiohttp==3.9.5
httpx==0.27.0
python-dotenv==1.0.1
structlog==24.1.0
